moved chat page to .../irc & starting to implement tab based chats
