+++ Multitool WinUI/Pages/Irc/IrcChatPage.xaml
@@ -1,0 +1,63 @@
+﻿<Page
+    x:Name="Page"
+    x:Class="MultitoolWinUI.Pages.IrcChatPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:controls="using:MultitoolWinUI.Controls"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="2*"/>
+            <ColumnDefinition Width="1*"/>
+        </Grid.ColumnDefinitions>
+
+        <Grid
+            Grid.Column="0">
+            <Grid.RowDefinitions>
+                <RowDefinition Height="70px"/>
+                <RowDefinition/>
+            </Grid.RowDefinitions>
+
+            <AutoSuggestBox
+                x:Name="UriTextBox"
+                PlaceholderText="Search query"
+                QuerySubmitted="UriTextBox_QuerySubmitted"
+                Grid.Row="0"/>
+
+            <WebView2
+                x:Name="PageWebView"
+                HorizontalAlignment="Stretch"
+                VerticalAlignment="Stretch"
+                Grid.Row="1"/>
+        </Grid>
+
+        <Grid
+            Grid.Column="1">
+            <Grid.RowDefinitions>
+                <RowDefinition Height="80px"/>
+                <RowDefinition Height="10*"/>
+            </Grid.RowDefinitions>
+
+            <TextBox
+                Text="{x:Bind Login, Mode=TwoWay}"
+                PlaceholderText="Login"
+                TextWrapping="Wrap"
+                HorizontalAlignment="Stretch"
+                VerticalAlignment="Center"
+                HorizontalContentAlignment="Center"
+                VerticalContentAlignment="Center"
+                BorderThickness="0"
+                Grid.Row="0"/>
+
+            <TabView
+                x:Name="Chats"
+                VerticalAlignment="Stretch"
+                Grid.Row="1"
+                AddTabButtonClick="Chats_AddTabButtonClick"
+                TabCloseRequested="Chats_TabCloseRequested">
+            </TabView>
+        </Grid>
+    </Grid>
+</Page>
