--- Multitool WinUI/Pages/IrcChatPage.xaml
@@ -1,147 +1,0 @@
-﻿<Page
-    x:Name="Page"
-    x:Class="MultitoolWinUI.Pages.IrcChatPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="2*"/>
-            <ColumnDefinition Width="1*"/>
-        </Grid.ColumnDefinitions>
-
-        <Grid
-            Grid.Column="0">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="70px"/>
-                <RowDefinition/>
-            </Grid.RowDefinitions>
-
-            <AutoSuggestBox
-                x:Name="UriTextBox"
-                PlaceholderText="Search query"
-                QuerySubmitted="UriTextBox_QuerySubmitted"
-                Grid.Row="0"/>
-
-            <WebView2
-                x:Name="PageWebView"
-                HorizontalAlignment="Stretch"
-                VerticalAlignment="Stretch"
-                Grid.Row="1"/>
-        </Grid>
-
-        <Grid
-            Grid.Column="1">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="80px"/>
-                <RowDefinition Height="10*"/>
-            </Grid.RowDefinitions>
-
-            <Grid
-                Margin="5"
-                ColumnSpacing="5"
-                Grid.Row="0">
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="2*"/>
-                    <ColumnDefinition Width="2*"/>
-                    <ColumnDefinition Width="1*"/>
-                    <ColumnDefinition Width="1*"/>
-                    <ColumnDefinition Width="8*"/>
-                </Grid.ColumnDefinitions>
-                <Grid.Resources>
-                    <Style TargetType="Button">
-                        <Setter Property="HorizontalAlignment" Value="Stretch"/>
-                        <Setter Property="VerticalAlignment" Value="Center"/>
-                    </Style>
-                    <Style TargetType="TextBox">
-                        <Setter Property="HorizontalAlignment" Value="Stretch"/>
-                        <Setter Property="HorizontalContentAlignment" Value="Center"/>
-                        <Setter Property="VerticalContentAlignment" Value="Center"/>
-                        <Setter Property="VerticalAlignment" Value="Center"/>
-                        <Setter Property="BorderThickness" Value="0"/>
-                    </Style>
-                    <Style TargetType="TextBlock">
-                        <Setter Property="HorizontalAlignment" Value="Stretch"/>
-                        <Setter Property="VerticalAlignment" Value="Center"/>
-                    </Style>
-                </Grid.Resources>
-                <Button
-                    x:Name="ClearChat"
-                    Content="Clear"
-                    Grid.Column="0"
-                    Click="ClearChat_Click"/>
-
-                <Button
-                    x:Name="LeaveChat"
-                    Grid.Column="1"
-                    Click="LeaveChat_Click">
-                    <TextBlock
-                        Text="Leave chat"
-                        TextWrapping="WrapWholeWords"/>
-                </Button>
-
-                <TextBlock
-                    x:Name="NoMTextBlock"
-                    Grid.Column="2"/>
-
-                <TextBlock
-                    x:Name="MPSTextBlock"
-                    Grid.Column="3"/>
-
-                <TextBox
-                    Text="{x:Bind Login, Mode=TwoWay}"
-                    PlaceholderText="Login"
-                    TextWrapping="Wrap"
-                    Grid.Column="4"/>
-            </Grid>
-
-            <ListView
-                x:Name="ChatListView"
-                ItemsSource="{x:Bind Chat}"
-                Background="{ThemeResource AppBarBorderThemeBrush}"
-                Grid.Row="1"
-                HorizontalAlignment="Stretch"
-                VerticalAlignment="Stretch">
-                <ListView.ItemsPanel>
-                    <ItemsPanelTemplate>
-                        <ItemsStackPanel
-                            VerticalAlignment="Bottom"
-                            ItemsUpdatingScrollMode="KeepLastItemInView"
-                            CacheLength="200"/>
-                    </ItemsPanelTemplate>
-                </ListView.ItemsPanel>
-
-                <ListView.ItemTemplate>
-                    <DataTemplate>
-
-                        <Grid Width="Auto" ColumnSpacing="3">
-                            <Grid.ColumnDefinitions>
-                                <ColumnDefinition Width="35px"/>
-                                <ColumnDefinition Width="1*"/>
-                                <ColumnDefinition Width="10*"/>
-                            </Grid.ColumnDefinitions>
-                            <TextBlock 
-                                Text="{Binding Timestamp, Mode=OneWay}"
-                                Opacity="0.8"
-                                Grid.Column="0"/>
-
-                            <TextBlock 
-                                Text="{Binding UserName, Mode=OneWay}"
-                                TextWrapping="Wrap"
-                                Grid.Column="1"/>
-
-                            <TextBlock 
-                                Text="{Binding Message, Mode=OneWay}"
-                                TextWrapping="WrapWholeWords"
-                                Grid.Column="2"/>
-                        </Grid>
-
-                    </DataTemplate>
-                </ListView.ItemTemplate>
-            </ListView>
-        </Grid>
-    </Grid>
-</Page>
